fix type

+++ hw1/cs285/policies/MLP_policy.py
@@ -86,14 +86,14 @@ class MLPPolicy(BasePolicy, nn.Module, metaclass=abc.ABCMeta):
             observation = obs[None]
 
         # TODO return the action that the policy prescribes
-        observation = torch.from_numpy(observation)
+        observation = torch.from_numpy(observation).to(torch.float32)
         observation = observation.to(ptu.device)
         return self.forward(observation)
 
     # update/train this policy
     def update(self, observations, actions, **kwargs):
         action_hat = self.get_action(observations)
-        actions = torch.from_numpy(actions)
+        actions = torch.from_numpy(actions).to(torch.float32)
         actions = actions.to(ptu.device)
         loss = self.loss_fn(action_hat, actions)
         self.optimizer.zero_grad()
